Converted floats to doubles

+++ slimemold_simulation.h
@@ -2,19 +2,19 @@
 #define SLIMEMOLD_SIMULATION_H
 
 struct Agent {
-    float direction;
-    float x;
-    float y;
+    double direction;
+    double x;
+    double y;
 };
 
 struct Map {
-    float *grid;
+    double *grid;
     int width;
     int height;
 };
 
 void simulate_step(struct Map map, struct Agent *agents,
-                    int nagents, float movement_speed, float trail_deposit_rate,
-                    float movement_noise, float turn_rate, float dispersion_rate,
-                    float evaporation_rate);
+                    int nagents, double movement_speed, double trail_deposit_rate,
+                    double movement_noise, double turn_rate, double dispersion_rate,
+                    double evaporation_rate);
 #endif
